+++ R/testthat-ui.R
@@ -21,9 +21,6 @@
 #'   folder.
 #' @param ... Additional arguments passed to [testthat::compare()] to
 #'   control specifics of comparison.
-#' @param user_fonts Passed to \code{svglite()} to make sure SVG are
-#'   reproducible. Defaults to Liberation fonts for standard families
-#'   and Symbola font for symbols.
 #' @param verbose If `TRUE`, the contents of the SVG files for the
 #'   comparison plots are printed during testthat checks. This is
 #'   useful to investigate errors when testing remotely.
@@ -32,6 +29,10 @@
 #'   interactive use. This is because there is no way of figuring out
 #'   in which directory this SVG lives. Consequently, only the test
 #'   case is printed.
+#' @param writer A function that takes the plot, a target SVG file,
+#'   and an optional plot title. It should transform the plot to SVG
+#'   in a deterministic way and write it to the target file. See
+#'   [write_svg()] (the default) for an example.
 #' @export
 #' @examples
 #' disp_hist_base <- function() hist(mtcars$disp)
@@ -40,11 +41,15 @@
 #' library("ggplot2")
 #' disp_hist_ggplot <- ggplot(mtcars, aes(disp)) + geom_histogram()
 #' expect_doppelganger("disp-histogram-ggplot", disp_hist_ggplot)
-expect_doppelganger <- function(title, fig, path = NULL, ...,
-                                user_fonts = NULL, verbose = FALSE) {
+expect_doppelganger <- function(title,
+                                fig,
+                                path = NULL,
+                                ...,
+                                verbose = FALSE,
+                                writer = write_svg) {
   fig_name <- str_standardise(title)
   testcase <- make_testcase_file(fig_name)
-  write_svg(fig, testcase, title, user_fonts)
+  writer(fig, testcase, title)
 
   context <- get(".context", envir = testthat::get_reporter())
   context <- str_standardise(context %||% "")