+++ tests/testthat/mock-pkg/tests/testthat/test-deps.R
@@ -1,7 +1,0 @@
-
-context("Dependencies")
-library("vdiffr")
-
-test_that("DESCRIPTION notes are updated manually", {
-  add_dependency("utils")
-})