+++ R/testthat-ui.R
@@ -137,7 +137,7 @@ compare_figs <- function(case) {
 # verbose cases at a later point.
 maybe_print_svgs <- function(case, pkg_path = NULL) {
   if (case$verbose && is_null(active_collecter())) {
-    cat_line(svg_files_lines(case, pkg_path))
+    cat_line(svg_files_lines(case, pkg_path), file = stderr())
   }
 }