+++ tests/testthat/test-mismatch.R
@@ -1,19 +1,18 @@
 context("mismatch")
 
 test_that("failures are pushed to log file", {
-  if (!is_checking()) {
-    expect_true(TRUE)
-    return("Skipping R CMD check tests")
+  log_path <- file.path(mock_pkg_dir, "tests", "vdiffr.Rout.fail")
+  if (!file.exists(log_path)) {
+    abort("Can't find log")
   }
+  log <- readLines(log_path)
 
-  log <- readLines(file.path(mock_pkg_dir, "tests", "vdiffr.Rout.fail"))
-
-  n_logged <- length(grep(">> Failed doppelganger:", log))
+  n_logged <- length(grep("Failed doppelganger:", log))
   n_failures <- length(keep(mock_cases, inherits, "mismatch_case"))
   expect_identical(n_logged, n_failures)
 })
 
 test_that("Verbose outputs svg files", {
   output <- mock_cases_outputs$output
-  expect_true(grepl(">> Failed doppelganger: myplot", output))
+  expect_true(grepl("Failed doppelganger: myplot", output))
 })