+++ R/testthat-ui.R
@@ -1,13 +1,19 @@
 #' Does a figure look like its expected output?
 #'
-#' If the test has never been validated yet, the test is skipped. If
-#' the test has previously been validated but \code{fig} does not look
-#' like its expected output, an error is issued. Use
-#' [validate_cases()] or [manage_cases()] to (re)validate
-#' the test.
+#' @description
 #'
-#' `fig` can be a ggplot object, a recordedplot, a function to be
-#' called, or more generally any object with a `print` method.
+#' `expect_doppelganger()` takes a figure to check visually.
+#'
+#' * If the figure has yet to be validated, the test is skipped. Call
+#'   [manage_cases()] to validate the new figure, so vdiffr knows what
+#'   to compare against.
+#'
+#' * If the test has been validated, `fig` is compared to the
+#'   validated figure. If the plot differs, a failure is issued
+#'   (except on CRAN, see section on regression testing below).
+#'
+#'   Either fix the problem, or call [manage_cases()] to validate the
+#'   new figure appearance.
 #'
 #' @param title A brief description of what is being tested in the
 #'   figure. For instance: "Points and lines overlap".
@@ -17,7 +23,12 @@
 #'
 #'   The title is also used as file name for storing SVG (in a
 #'   sanitzed form, with special characters converted to `"-"`).
-#' @param fig A figure to test.
+#' @param fig A figure to test. This can be a ggplot object, a
+#'   recordedplot, or more generally any object with a `print` method.
+#'
+#'   For plots that can't be represented as printable objects, you can
+#'   pass a function. This function must construct the plot and print
+#'   it.
 #' @param path The path where the test case should be stored, relative
 #'   to the `tests/figs/` folder. If `NULL` (the default), the current
 #'   testthat context is used to create a subfolder. Supply an empty
@@ -30,6 +41,31 @@
 #'   and an optional plot title. It should transform the plot to SVG
 #'   in a deterministic way and write it to the target file. See
 #'   [write_svg()] (the default) for an example.
+#'
+#' @section Regression testing versus Unit testing:
+#'
+#' Failures to match a validated appearance are only reported when the
+#' tests are run locally, on Travis, Appveyor, or any environment
+#' where the `Sys.getenv("CI")` or `Sys.getenv("NOT_CRAN")` variables
+#' are set. Because vdiffr is more of a monitoring than a unit testing
+#' tool, it shouldn't cause R CMD check failures on the CRAN machines.
+#'
+#' Checking the appearance of a figure is inherently fragile. It is
+#' similar to testing for errors by matching exact error messages:
+#' these messages are susceptible to change at any time. Similarly,
+#' the appearance of plots depends on a lot of upstream code, such as
+#' the way margins and spacing are computed. vdiffr uses a special
+#' ggplot2 theme that should change very rarely, but there are just
+#' too many upstream factors that could cause breakages. For this
+#' reason, figure mismatches are not necessarily representative of
+#' actual failures.
+#'
+#' Visual testing is not an alternative to writing unit tests for the
+#' internal data transformations performed during the creation of your
+#' figure. It is more of a monitoring tool that allows you to quickly
+#' check how the appearance of your figures changes over time, and to
+#' manually assess whether changes reflect actual problems in your
+#' package.
 #'
 #' @section Debugging:
 #'
@@ -167,6 +203,7 @@ new_expectation <- function(msg, case, type, vdiffr_type) {
   classes <- c(class(exp), vdiffr_type)
   structure(exp, class = classes, vdiffr_case = case)
 }
+
 new_exp <- function(msg, case) {
   new_expectation(msg, case, "skip", "vdiffr_new")
 }
@@ -177,8 +214,10 @@ mismatch_exp <- function(msg, case) {
   if (is_vdiffr_stale()) {
     msg <- "The vdiffr engine is too old. Please update vdiffr and revalidate the figures."
     new_expectation(msg, case, "skip", "vdiffr_mismatch")
-  } else {
+  } else if (is_ci()) {
     new_expectation(msg, case, "failure", "vdiffr_mismatch")
+  } else {
+    new_expectation(msg, case, "skip", "vdiffr_mismatch")
   }
 }