+++ NEWS.md
@@ -1,6 +1,9 @@
 
 # vdiffr 0.2.3.9000
 
+* `write_svg()` is now an exported function. It provides a template
+  (function arguments and return value) for SVG writer functions.
+
 * The `user_fonts` argument of `expect_doppelganger()` is defunct. It
   wasn't used in any package and complicated the UI for no real
   benefit. The fonts used to generate the SVGs are now hardcoded to