+++ NEWS.md
@@ -1,6 +1,12 @@
 
 # vdiffr 0.2.3.9000
 
+* `expect_doppelganger()` now takes a `writer` argument. This makes it
+  easy to use vdiffr with a different SVG engine. See `?write_svg` for
+  an example function. Packages implementing a different SVG engine
+  should wrap around `expect_doppelganger()` to pass their custom
+  writer.
+
 * `write_svg()` is now an exported function. It provides a template
   (function arguments and return value) for SVG writer functions.