+++ NEWS.md
@@ -1,5 +1,5 @@
 
-# vdiffr 0.2.99.9000
+# vdiffr 0.3.0
 
 This release of vdiffr features a major overhaul of the internals to
 make the package more robust.